requirements.txt
python-telegram-bot==20.4
requests==2.31.0
pylint==2.17.5
pep8==1.7.1
schedule==1.2.0